pyproject.toml
[project]
name = "ragproductionapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.73.1",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.20",
    "llama-index-embeddings-voyageai>=0.6.0",
    "llama-index-readers-file>=0.6.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
    "voyageai>=0.3.7",
]
